Move cube load and save functions to esmvaltool.preprocessor._io.py
